fix: address unresolved review followups
