+++ .github/dependabot.yml
@@ -6,7 +6,7 @@ updates:
       interval: weekly
     open-pull-requests-limit: 5
 
-  - package-ecosystem: npm
+  - package-ecosystem: bun
     directory: /panel
     schedule:
       interval: weekly
